removed the method updateToUser() and any calls to it
in the call on 28 Oct 2024, Kunala, Isaac, and Shawn agreed that since UserRepository and AccountRegistrationRepository (not available on this branch, but on T46 instead) have the same logic of removing an approved request and making a user out of it, it would be more appropriate to have the latter hold the code instead. So the code is being removed from the former.