error fixing
some events were crashing upon inquiry of canDeleteEvent(Event). This would happen if there were no approvedRequests for that event and thus that member would be read from fb as null. Then the method canDeleteEvent would call a method on that null object, therefore causing a crash. This commit fixes it.